requirements.txt
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
arxiv==3.0.0
boto3==1.42.95
botocore==1.42.95
certifi==2026.4.22
charset-normalizer==3.4.7
distro==1.9.0
docstring_parser==0.18.0
feedparser==6.0.12
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.3.2
langchain-protocol==0.0.11
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.35
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
playwright==1.58.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
PyMuPDF==1.27.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
s3transfer==0.16.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
